requirements.txt
shapely
pyyaml